requirements.txt
awsebcli==3.17.1
boto==2.49.0
boto3==1.12.0
botocore==1.14.17
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
Django==3.0.3
django-crispy-forms==1.8.1
django-ses==0.8.14
django-storages==1.9.1
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
pathspec==0.5.9
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.2
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.7
